chore(api): harden integration test infrastructure
Three small fixes uncovered while bringing up the ADR-002 invariant
suite against a real Postgres:

* `env.ts`: default `DATABASE_URL_TEST` now targets the new
  `postgres_test` container (port 5434) and appends
  `connection_limit=20` — the property test runs up to 10 parallel
  openPeriod transactions each holding a connection through ~5s lock
  waits, which exhausts the default Prisma pool (~9) and surfaces as
  "Unable to start a transaction in the given time" instead of the
  expected PERIOD_LOCK_TIMEOUT.

* `env.ts`: `FAST_CHECK_RUNS` default lowered from 100 (local) to 25.
  Concurrency-heavy property tests do TRUNCATE + factories + N
  parallel ops per run; 100 runs blow past any reasonable testTimeout.
  25 is sufficient for the shrinker to find counterexamples.

* `jest-integration.json`: `testTimeout` raised from 30s to 120s for
  the same reason — a single property assertion can legitimately take
  60–90s with worst-case lock waits.

* `arbitraries.ts`: `fc.float({min:0.01})` failed compile in
  fast-check 3.x ("min must be a 32-bit float"). Switched to
  `fc.double` — 64-bit, matches Decimal-money semantics anyway.

+++ apps/api/test/integration/fixtures/arbitraries.ts
@@ -1,10 +1,12 @@
 // apps/api/test/integration/fixtures/arbitraries.ts
 import * as fc from 'fast-check';
 
-// BoQ shape — N items, each with a positive contract_value
+// BoQ shape — N items, each with a positive contract_value.
+// fc.double (64-bit) — fc.float requires 32-bit representable bounds,
+// and money/Decimal semantics need double precision anyway.
 export const arbBoQShape = fc.record({
   contractValues: fc.array(
-    fc.float({ min: 0.01, max: 10_000, noNaN: true, noDefaultInfinity: true }),
+    fc.double({ min: 0.01, max: 10_000, noNaN: true, noDefaultInfinity: true }),
     { minLength: 2, maxLength: 50 },
   ),
 });
